fix: 클라이언트 컴포넌트에서 index.ts를 import하지 않게
