+++ apis/index.ts
@@ -2,6 +2,8 @@ import { cookies } from 'next/headers';
 
 import { objToQueryString } from '@/utils/convertParams';
 
+import { NetworkError, checkError } from './error';
+
 export const BASE_URL = 'https://cse-dev-waffle.bacchus.io/api/v1';
 
 // REST request
@@ -33,8 +35,6 @@ export const deleteRequest = async (url: string, init?: RequestInit) => {
   await fetchWithRetry(`${BASE_URL}${url}`, 'DELETE', init);
 };
 
-// util
-
 /**
  * 학교 서버가 간헐적으로 첫 요청에서 ECONNRESET이 뜨는 경우가 있어 한 번 더 요청.
  * 안전을 위해 멱등성이 있는 GET 요청에 대해서만 retry
@@ -46,7 +46,8 @@ const fetchWithRetry = async (url: string, method: string, init?: RequestInit) =
     return await fetchWithCredentials(url, method, init);
   } catch (e) {
     if (e instanceof NetworkError) throw e;
-    console.log('fetchWithRetry: 서버 에러 감지');
+
+    console.error(`fetchWithRetry: ${e}`);
     return fetchWithCredentials(url, method, init);
   }
 };
@@ -65,16 +66,3 @@ const fetchWithCredentials = async (url: string, method: string, init?: RequestI
 
   return resp;
 };
-
-export const checkError = (response: Response) => {
-  if (response.ok) return;
-  throw new NetworkError(response.status);
-};
-
-export class NetworkError extends Error {
-  statusCode: number;
-  constructor(statusCode: number) {
-    super(`${statusCode} 에러`);
-    this.statusCode = statusCode;
-  }
-}
