+++ components/reservations/modals/useAddReservation.ts
@@ -1,7 +1,8 @@
 import { useReducer, useState, FormEventHandler } from 'react';
 
 import { postReservation } from '@/actions/reservation';
-import { NetworkError } from '@/apis';
+
+import { NetworkError } from '@/apis/error';
 
 import { ReservationPostBody } from '@/types/reservation';
